Remove Redis: simplify stack
Core changes:
* Use `redis-mock` package, see https://www.npmjs.com/package/redis-mock
or https://github.com/yeahoffline/redis-mock. This allows us to remove redis
as a dependency as this package does all operations in-memory, as per the
documentation "All operations are performed in-memory, so no Redis installation is required.".
* Most of the logic sits in the file: `services/ob-api-proxy/app/session/persistence.js`.
* Tidyup `services/ob-api-proxy/app/session/persistence.js`.

Result of core changes:
* Remove Redis as we are using a mock now.
* Remove Redis configuration environmental variables.
* Remove `redis` from `docker-compose.yml`.
* Remove `redis` from `package.json`.
* Remove Redis anywhere grep finds it.

Pending:
* Remove logic around persisting the session, i.e., remove the
`services/ob-api-proxy/app/session/persistence.js` and rewrite logic
that calls into it.

---

REFAPP-411
Remove Redis
https://openbanking.atlassian.net/browse/REFAPP-411