Merged in bugfix/REFAPP-638-missing-issuer-url-in-config (pull request #207)
REFAPP-638: Missing "issuer" URL in config

* `issuer`: fetch from `/.well-known/openid-configuration`

Add issuer URL to config. The config should have an option to for the issuer to be pulled from the well-known which is passed back to the config.

---
REFAPP-638
Missing "issuer" URL in config

* `issuer`: fetch from `/.well-known/openid-configuration`

Add issuer URL to config. The config should have an option to for the issuer to be pulled from the well-known which is passed back to the config.

---
REFAPP-638
Missing "issuer" URL in config

Approved-by: Glyn Jackson
Approved-by: José Pinto