Signature Configuration Refactor
Previously the tool's confguration tried to cater for different signature
setups with multiple conditional variables. This change reduces the
ambiguity by requiring only the minimum number
of fileds that's
needed. This makes the tool more flexible with the tradeoff that users
need to know exactly how the TPP signature claims are populated.